change reservation modal display, bug fix, and organize file
