+++ src/core/components/pages/Reservation/MyReservationPage.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { Route, Switch, useRouteMatch } from "react-router-dom"
 import ReservationPage from "../Reservation/Reservation"
-import ReservationDetail from "../Reservation/ReservationDetail"
+import ReservationDetail from "./ReservationDetail/ReservationDetailPage"
 import withUserGuard from "../../../guards/user.guard"
 
 const MyReservationPage = () => {
