+++ src/locales/th/translation.json
@@ -318,5 +318,7 @@
   "paymentApproved": "หลักฐานการชำระเงินได้รับการอนุมัติแล้ว",
   "noEmpty": "❌ กรุณากรอกข้อมูล",
   "landingMessage": "จองสนามกีฬาออนไลน์และเช็คอินได้ง่ายๆในที่เดียว",
-  "fillHere": "กรอกข้อมูล"
+  "fillHere": "กรอกข้อมูล",
+  "reminder": "หมายเหตุ",
+  "cancellationWarning": "หากทำการยกเลิกการจองตอนนี้ คุณและเพื่อนในทีมจะถูกแบนเป็นเวลา {{lateCancellationPunishment}} วัน"
 }
